Many edits to fix camera issues
There was an issue with the `raspivid | ffmpeg` process when spawned from Node.  I'm not sure
why yet, but when these processes are spawned from Node they will hang after a few minutes.
Running them from the `run.sh` script seems to work fine indefinitely.

I also modified the program to encode the video stream as MJPEG instead of h.264, which seems
to optimize the write efficiency. Writing to the filesystem also seems to work, but I improved
it further by writing to /dev/shm, which reduces read/write latencies.

I still get bad pictures on some frames, an issue I'll need to address later, but for now this
seems to work rather well and I get a higher FPS.

I also modified the write operations to queue up the images and write in large batches, which
improves the system performance.